requirements.txt
ecdsa==0.11
pycrypto==2.6.1
